+++ internal/models/balance/balance.go
@@ -3,17 +3,17 @@ package balance
 import "time"
 
 type Response struct {
-	Current   float64 `json:"current"`
-	Withdrawn float64 `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
 
 type WithdrawRequest struct {
 	Order string  `json:"order"`
-	Sum   float64 `json:"sum"`
+	Sum   float32 `json:"sum"`
 }
 
 type Withdrawal struct {
 	ProcessedAt time.Time `json:"processed_at"`
 	Order       string    `json:"order"`
-	Sum         float64   `json:"sum"`
+	Sum         float32   `json:"sum"`
 }
